entity modify

+++ dsm-hackathon-backend/src/main/kotlin/org/dsmhackathonbackend/domain/company/domain/Company.kt
@@ -18,7 +18,7 @@ class Company(
 
     var info: String,
 
-    var rating: Int,
+    var rating: Float,
 
     val registrationNumber: Int,
